Cleanup, merge modules that ended up not growing
