--- src/st7701s_spi/instructions/mod.rs
@@ -1,1 +1,0 @@
-pub mod power;
