--- src/st7701s_spi/instructions/power.rs
@@ -1,30 +1,0 @@
-use crate::{ bit_value_enum};
-
-bit_value_enum! {
-    pub enum BiasCurrent<2> {
-        #[default]
-        const Off = 0b00,
-        const Min = 0b01,
-        const Mid = 0b10,
-        const Max = 0b11,
-    }
-}
-
-// transmission_mapping! {
-//     /// ## Power Control 1
-//     /// Primary power supply configuration
-//     /// > Reference: `PWCTRL1` p. 292
-//     pub struct BiasCurrentParameters<1>(
-//         0: (
-//             D6(gamma_bias<2> as BiasCurrent = Mid),
-//             D2(source_input_bias<2> as BiasCurrent = Max),
-//             D0(source_output_bias<2> as BiasCurrent = Off),
-//         ),
-//     );
-// }
-
-// impl<C: Connection> ST7701S<C, ExtensionBk1> {
-//     pub fn select_bias_current(&self, transmission: &BiasCurrentParameters) -> InstructionResult {
-//         self.write::<PWCTRL1>(transmission)
-//     }
-// }
