+++ src/st7701s_spi/parameters/bk1_power.rs
@@ -9,9 +9,6 @@ use crate::{
 use MipiLaneCount::OneLane;
 use Switch::{On, Off};
 
-// ============================================================================
-// MIPI Settings (MIPISET1-4) - p. 299-302
-// ============================================================================
 
 bit_value_enum! {
     /// MIPI Lane Configuration
@@ -95,9 +92,6 @@ transmission_mapping! {
     );
 }
 
-// ============================================================================
-// Voltage Control Settings - p. 287-291
-// ============================================================================
 
 transmission_mapping! {
     pub struct OperatingVoltage<1>(
@@ -313,9 +307,6 @@ impl Display for GateLowVoltage {
     }
 }
 
-// ============================================================================
-// Power Control Settings - p. 292-293
-// ============================================================================
 
 bit_value_enum! {
     /// VGH Voltage Multiplier
@@ -351,9 +342,6 @@ transmission_mapping! {
     );
 }
 
-// ============================================================================
-// Panel Clock Settings - p. 294-296
-// ============================================================================
 
 bit_value_enum! {
     /// Clock Divider Ratio
@@ -400,9 +388,6 @@ transmission_mapping! {
     );
 }
 
-// ============================================================================
-// Source Drive Timing - p. 297-298
-// ============================================================================
 
 bit_value_enum! {
     /// Source Pre-charge Period
@@ -440,9 +425,6 @@ transmission_mapping! {
     );
 }
 
-// ============================================================================
-// Test Command - p. 290
-// ============================================================================
 
 /// Test Command Parameter
 /// Used for factory testing and diagnostics
