+++ src/st7701s_spi/parameters/brightness.rs
@@ -5,9 +5,6 @@ use crate::{
 };
 use Switch::Off;
 
-// ============================================================================
-// Display Brightness - p. 221
-// ============================================================================
 
 transmission_mapping!(
     /// ## Display Brightness Value
@@ -18,9 +15,6 @@ transmission_mapping!(
     );
 );
 
-// ============================================================================
-// Brightness Control - p. 223
-// ============================================================================
 
 transmission_mapping! {
     /// ## Display Brightness Control Modes
@@ -35,9 +29,6 @@ transmission_mapping! {
     );
 }
 
-// ============================================================================
-// Content Adaptive Brightness Control (CABC) - p. 225, 227
-// ============================================================================
 
 bit_value_enum! {
     /// CABC Mode Selection
@@ -62,9 +53,6 @@ transmission_mapping! {
     );
 }
 
-// ============================================================================
-// CABC Minimum Brightness - p. 229, 230
-// ============================================================================
 
 transmission_mapping! {
     /// ## CABC Minimum Brightness
