+++ src/st7701s_spi/parameters/color.rs
@@ -1,8 +1,5 @@
 use crate::bit_value_enum;
 
-// ============================================================================
-// Pixel Format
-// ============================================================================
 
 bit_value_enum! {
     /// Bits Per Pixel Format
@@ -14,9 +11,6 @@ bit_value_enum! {
     }
 }
 
-// ============================================================================
-// Color Enhancement
-// ============================================================================
 
 bit_value_enum! {
     /// Color Enhancement Level
@@ -28,9 +22,6 @@ bit_value_enum! {
     }
 }
 
-// ============================================================================
-// Color Channels and Pixel States
-// ============================================================================
 
 /// Individual Color Channel
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
