+++ src/st7701s_spi/parameters/display.rs
@@ -5,9 +5,6 @@ use crate::{
 };
 use Switch::On;
 
-// ============================================================================
-// Tearing Effect - p. 212
-// ============================================================================
 
 bit_value_enum! {
     /// Tearing Effect Signal Mode
@@ -27,9 +24,6 @@ transmission_mapping! {
     );
 }
 
-// ============================================================================
-// Gamma Curve Selection - p. 208
-// ============================================================================
 
 // FIXME: This isn't necessarily correct, and the datasheet indicates
 // that the initial value is "RESERVED". Maybe better to set this to
@@ -55,9 +49,6 @@ transmission_mapping! {
     );
 }
 
-// ============================================================================
-// Gamma Voltage Control - p. 261-264
-// ============================================================================
 
 bit_value_enum! {
     /// Voltage Bias Adjustment
@@ -96,9 +87,6 @@ transmission_mapping! {
     );
 }
 
-// ============================================================================
-// Display Image Mode - p. 204-208
-// ============================================================================
 
 transmission_mapping! {
     /// ## Display Image Mode Settings
@@ -114,9 +102,6 @@ transmission_mapping! {
     );
 }
 
-// ============================================================================
-// Display Signal Mode - p. 211-212
-// ============================================================================
 
 transmission_mapping! {
     /// ## Tearing Effect Signal Configuration
@@ -130,9 +115,6 @@ transmission_mapping! {
     );
 }
 
-// ============================================================================
-// Line Settings - p. 213
-// ============================================================================
 
 transmission_mapping!(
     /// ## Tearing Effect Scan Line
@@ -144,9 +126,6 @@ transmission_mapping!(
     );
 );
 
-// ============================================================================
-// Porch Control - p. 269
-// ============================================================================
 
 transmission_mapping!(
     /// ## Vertical Porch Control
@@ -158,9 +137,6 @@ transmission_mapping!(
     );
 );
 
-// ============================================================================
-// Inversion Selection - p. 270
-// ============================================================================
 
 bit_value_enum!(
     /// Polarity Inversion Pattern
