+++ src/st7701s_spi/parameters/power.rs
@@ -10,3 +10,12 @@ state_struct! {
   }
 }
 
+crate::bit_value_enum! {
+    pub enum BiasCurrent<2> {
+        #[default]
+        const Off = 0b00,
+        const Min = 0b01,
+        const Mid = 0b10,
+        const Max = 0b11,
+    }
+}
