--- src/st7701s_spi/state/derived.rs
@@ -1,10 +1,0 @@
-#[derive(Debug, Hash, Copy, Clone, PartialEq, Eq)]
-pub enum Power {
-    L1,
-    L2,
-    L3,
-    L4,
-    L5,
-    L6,
-    L7,
-}
