+++ src/st7701s_spi/state/mod.rs
@@ -1,3 +1,2 @@
 pub mod abstractions;
-pub mod derived;
 pub mod domains;
